libuv cares: scope manual localhost resolution to only Windows
